Separate the logic into ConsensusCommitUtils
