+++ core/src/test/java/com/scalar/db/transaction/consensuscommit/ConsensusCommitAdminTest.java
@@ -273,150 +273,4 @@ public class ConsensusCommitAdminTest {
                         .build()))
         .isInstanceOf(IllegalArgumentException.class);
   }
-
-  @Test
-  public void
-      isTransactionalTableMetadata_properTransactionalTableMetadataGiven_shouldReturnTrue() {
-    // Arrange
-    final String ACCOUNT_ID = "account_id";
-    final String ACCOUNT_TYPE = "account_type";
-    final String BALANCE = "balance";
-
-    TableMetadata metadata =
-        TableMetadata.newBuilder()
-            .addColumn(ACCOUNT_ID, DataType.INT)
-            .addColumn(ACCOUNT_TYPE, DataType.INT)
-            .addColumn(BALANCE, DataType.INT)
-            .addColumn(Attribute.ID, DataType.TEXT)
-            .addColumn(Attribute.STATE, DataType.INT)
-            .addColumn(Attribute.VERSION, DataType.INT)
-            .addColumn(Attribute.PREPARED_AT, DataType.BIGINT)
-            .addColumn(Attribute.COMMITTED_AT, DataType.BIGINT)
-            .addColumn(Attribute.BEFORE_PREFIX + BALANCE, DataType.INT)
-            .addColumn(Attribute.BEFORE_ID, DataType.TEXT)
-            .addColumn(Attribute.BEFORE_STATE, DataType.INT)
-            .addColumn(Attribute.BEFORE_VERSION, DataType.INT)
-            .addColumn(Attribute.BEFORE_PREPARED_AT, DataType.BIGINT)
-            .addColumn(Attribute.BEFORE_COMMITTED_AT, DataType.BIGINT)
-            .addPartitionKey(ACCOUNT_ID)
-            .addClusteringKey(ACCOUNT_TYPE)
-            .build();
-
-    // Act
-    boolean actual = ConsensusCommitAdmin.isTransactionalTableMetadata(metadata);
-
-    // Assert
-    assertThat(actual).isTrue();
-  }
-
-  @Test
-  public void
-      isTransactionalTableMetadata_transactionalTableMetadataWithoutMetaColumnGiven_shouldReturnFalse() {
-    // Arrange
-    final String ACCOUNT_ID = "account_id";
-    final String ACCOUNT_TYPE = "account_type";
-    final String BALANCE = "balance";
-
-    TableMetadata metadata =
-        TableMetadata.newBuilder()
-            .addColumn(ACCOUNT_ID, DataType.INT)
-            .addColumn(ACCOUNT_TYPE, DataType.INT)
-            .addColumn(BALANCE, DataType.INT)
-            .addColumn(Attribute.STATE, DataType.INT)
-            .addColumn(Attribute.VERSION, DataType.INT)
-            .addColumn(Attribute.PREPARED_AT, DataType.BIGINT)
-            .addColumn(Attribute.COMMITTED_AT, DataType.BIGINT)
-            .addColumn(Attribute.BEFORE_PREFIX + BALANCE, DataType.INT)
-            .addColumn(Attribute.BEFORE_ID, DataType.TEXT)
-            .addColumn(Attribute.BEFORE_STATE, DataType.INT)
-            .addColumn(Attribute.BEFORE_VERSION, DataType.INT)
-            .addColumn(Attribute.BEFORE_PREPARED_AT, DataType.BIGINT)
-            .addColumn(Attribute.BEFORE_COMMITTED_AT, DataType.BIGINT)
-            .addPartitionKey(ACCOUNT_ID)
-            .addClusteringKey(ACCOUNT_TYPE)
-            .build();
-
-    // Act
-    boolean actual = ConsensusCommitAdmin.isTransactionalTableMetadata(metadata);
-
-    // Assert
-    assertThat(actual).isFalse();
-  }
-
-  @Test
-  public void
-      isTransactionalTableMetadata_transactionalTableMetadataWithoutProperBeforeColumnGiven_shouldReturnFalse() {
-    // Arrange
-    final String ACCOUNT_ID = "account_id";
-    final String ACCOUNT_TYPE = "account_type";
-    final String BALANCE = "balance";
-
-    TableMetadata metadata =
-        TableMetadata.newBuilder()
-            .addColumn(ACCOUNT_ID, DataType.INT)
-            .addColumn(ACCOUNT_TYPE, DataType.INT)
-            .addColumn(BALANCE, DataType.INT)
-            .addColumn(Attribute.ID, DataType.TEXT)
-            .addColumn(Attribute.STATE, DataType.INT)
-            .addColumn(Attribute.VERSION, DataType.INT)
-            .addColumn(Attribute.PREPARED_AT, DataType.BIGINT)
-            .addColumn(Attribute.COMMITTED_AT, DataType.BIGINT)
-            .addColumn(Attribute.BEFORE_ID, DataType.TEXT)
-            .addColumn(Attribute.BEFORE_STATE, DataType.INT)
-            .addColumn(Attribute.BEFORE_VERSION, DataType.INT)
-            .addColumn(Attribute.BEFORE_PREPARED_AT, DataType.BIGINT)
-            .addColumn(Attribute.BEFORE_COMMITTED_AT, DataType.BIGINT)
-            .addPartitionKey(ACCOUNT_ID)
-            .addClusteringKey(ACCOUNT_TYPE)
-            .build();
-
-    // Act
-    boolean actual = ConsensusCommitAdmin.isTransactionalTableMetadata(metadata);
-
-    // Assert
-    assertThat(actual).isFalse();
-  }
-
-  @Test
-  public void
-      removeTransactionalMetaColumns_transactionalTableMetadataWithoutProperBeforeColumnGiven_shouldReturnFalse() {
-    // Arrange
-    final String ACCOUNT_ID = "account_id";
-    final String ACCOUNT_TYPE = "account_type";
-    final String BALANCE = "balance";
-
-    TableMetadata metadata =
-        TableMetadata.newBuilder()
-            .addColumn(ACCOUNT_ID, DataType.INT)
-            .addColumn(ACCOUNT_TYPE, DataType.INT)
-            .addColumn(BALANCE, DataType.INT)
-            .addColumn(Attribute.ID, DataType.TEXT)
-            .addColumn(Attribute.STATE, DataType.INT)
-            .addColumn(Attribute.VERSION, DataType.INT)
-            .addColumn(Attribute.PREPARED_AT, DataType.BIGINT)
-            .addColumn(Attribute.COMMITTED_AT, DataType.BIGINT)
-            .addColumn(Attribute.BEFORE_PREFIX + BALANCE, DataType.INT)
-            .addColumn(Attribute.BEFORE_ID, DataType.TEXT)
-            .addColumn(Attribute.BEFORE_STATE, DataType.INT)
-            .addColumn(Attribute.BEFORE_VERSION, DataType.INT)
-            .addColumn(Attribute.BEFORE_PREPARED_AT, DataType.BIGINT)
-            .addColumn(Attribute.BEFORE_COMMITTED_AT, DataType.BIGINT)
-            .addPartitionKey(ACCOUNT_ID)
-            .addClusteringKey(ACCOUNT_TYPE)
-            .build();
-
-    // Act
-    TableMetadata actual = ConsensusCommitAdmin.removeTransactionalMetaColumns(metadata);
-
-    // Assert
-    assertThat(actual)
-        .isEqualTo(
-            TableMetadata.newBuilder()
-                .addColumn(ACCOUNT_ID, DataType.INT)
-                .addColumn(ACCOUNT_TYPE, DataType.INT)
-                .addColumn(BALANCE, DataType.INT)
-                .addPartitionKey(ACCOUNT_ID)
-                .addClusteringKey(ACCOUNT_TYPE)
-                .build());
-  }
 }
